React-Django-MySQL/server/requirement.txt
django-admin startproject server
cd server
python manage.py startapp api
pip install djangorestframework
python manage.py makemigrations # database
python manage.py migrate # database

python manage.py createsuperuser 
python manage.py runserver

class Product(models.Model):
    image = models.ImageField(upload_to='upload/images', null=False, blank=False)
    name = models.CharField(max_length=150, null=False, blank=False)
    price = models.DecimalField(max_digits=7, decimal_places=2, null=False, blank=False)
    decription = models.TextField()
    category = models.CharField(max_length=50, null=False, blank=True)

    def __str__(self) -> str:
        return super().__str__()